feat(auth-refactor): move more functions to wallet saga file
